Fix some recent test breakage

+++ src/test/java/edu/illinois/library/cantaloupe/resource/iiif/v3/ImageInfoFactoryTest.java
@@ -202,7 +202,7 @@ public class ImageInfoFactoryTest extends BaseTest {
         instance.setMaxScale(1);
 
         ImageInfo<String, Object> imageInfo = invokeNewImageInfo();
-        assertEquals(Math.round(1500 * 1200), imageInfo.get("maxArea"));
+        assertEquals((long) Math.round(1500 * 1200), imageInfo.get("maxArea"));
     }
 
     @Test
